Added the possibility to interrupt tasks to the concurrent library. Added a class that observes the running tasks and interrupts them, if they are waiting for a given time. Increased the version to 0.0.2-SNAPSHOT.

+++ topicmodeling.concurrent/src/test/java/org/aksw/simba/topicmodeling/concurrent/utils/MaxDurationCheckingThread.java
@@ -29,11 +29,11 @@ public class MaxDurationCheckingThread implements Runnable {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(MaxDurationCheckingThread.class);
 
-    private int duration;
+    private long duration;
     private boolean finished;
     private Thread watchedThread;
 
-    public MaxDurationCheckingThread(Thread watchedThread, int duration) {
+    public MaxDurationCheckingThread(Thread watchedThread, long duration) {
         this.duration = duration;
         this.watchedThread = watchedThread;
     }
